# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PrintFlow API - Moran Creative",
        "description": "API para gestión de pedidos de impresión 3D",
        "contact": {
            "email": "admin@morancreative.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/finishes": {
            "get": {
                "tags": [
                    "Acabados"
                ],
                "summary": "Listar acabados disponibles (público)",
                "operationId": "f04a5d2ebd5720414f1092398027c1ad",
                "responses": {
                    "200": {
                        "description": "Lista de acabados con su costo"
                    }
                }
            },
            "post": {
                "tags": [
                    "Acabados - Admin"
                ],
                "summary": "Crear acabado (solo admin)",
                "operationId": "f075679fd2322c22c23deb2562ca6893",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "fixed_cost"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Lijado fino"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Superficie suave al tacto"
                                    },
                                    "fixed_cost": {
                                        "type": "number",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Acabado creado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/finishes/{id}": {
            "put": {
                "tags": [
                    "Acabados - Admin"
                ],
                "summary": "Actualizar acabado (solo admin)",
                "operationId": "898909e786d9b34d2925dae79eecf120",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Lijado medio"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "fixed_cost": {
                                        "type": "number",
                                        "example": 40
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Acabado actualizado"
                    },
                    "404": {
                        "description": "Acabado no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Acabados - Admin"
                ],
                "summary": "Eliminar acabado (solo admin)",
                "description": "No se puede eliminar si hay pedidos que usan este acabado.",
                "operationId": "0245747b7d7e25e1cb414f69bfd93fc4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Acabado eliminado"
                    },
                    "422": {
                        "description": "No se puede eliminar, está en uso"
                    },
                    "404": {
                        "description": "Acabado no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}